+++ db2/history2/ledger_key.go
@@ -1,0 +1,41 @@
+package history2
+
+import (
+	"database/sql/driver"
+	"encoding/json"
+
+	"gitlab.com/distributed_lab/logan/v3/errors"
+	"gitlab.com/tokend/go/xdr"
+)
+
+type LedgerKey xdr.LedgerKey
+
+//Value - implements db driver method for auto marshal
+func (r LedgerKey) Value() (driver.Value, error) {
+	result, err := json.Marshal(r)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to marshal poll data")
+	}
+
+	return result, nil
+}
+
+//Scan - implements db driver method for auto unmarshal
+func (r *LedgerKey) Scan(src interface{}) error {
+	var data []byte
+	switch rawData := src.(type) {
+	case []byte:
+		data = rawData
+	case string:
+		data = []byte(rawData)
+	default:
+		return errors.New("Unexpected type for jsonb")
+	}
+
+	err := json.Unmarshal(data, r)
+	if err != nil {
+		return errors.Wrap(err, "failed to unmarshal ledger key")
+	}
+
+	return nil
+}