+++ ingest2/main.go
@@ -4,4 +4,4 @@
 package ingest2
 
 // increase it if you want force reingest (after backward not compatible changes)
-const CurrentIngestVersion = 0
+const CurrentIngestVersion = 1