+++ ingest2/operations/handler.go
@@ -130,6 +130,9 @@ func NewOperationsHandler(operationsStorage operationsStorage, participantEffect
 			xdr.OperationTypeManageAccountRule: &manageAccountRuleOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},
+			xdr.OperationTypeManageAccountSpecificRule: &manageAccountSpecificRuleOpHandler{
+				manageOfferOpHandler: manageOfferOpHandlerInst,
+			},
 			xdr.OperationTypeManageSigner: &manageSignerOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},