+++ ingest2/storage/sale.go
@@ -27,12 +27,12 @@ func (q *Sale) Insert(sale history2.Sale) error {
 		Columns(
 			"id", "owner_address", "base_asset", "default_quote_asset", "start_time", "end_time",
 			"quote_assets", "soft_cap", "hard_cap", "details", "base_current_cap",
-			"base_hard_cap, sale_type, state",
+			"base_hard_cap", "sale_type", "state", "version",
 		).
 		Values(
 			sale.ID, sale.OwnerAddress, sale.BaseAsset, sale.DefaultQuoteAsset, sale.StartTime, sale.EndTime,
 			sale.QuoteAssets, sale.SoftCap, sale.HardCap, sale.Details,
-			sale.BaseCurrentCap, sale.BaseHardCap, sale.SaleType, sale.State,
+			sale.BaseCurrentCap, sale.BaseHardCap, sale.SaleType, sale.State, sale.Version,
 		)
 
 	_, err := q.repo.Exec(sql)
@@ -59,6 +59,7 @@ func (q *Sale) Update(sale history2.Sale) error {
 		"base_current_cap":    sale.BaseCurrentCap,
 		"sale_type":           sale.SaleType,
 		"state":               sale.State,
+		"version":             sale.Version,
 	}).Where("id = ?", sale.ID)
 
 	_, err := q.repo.Exec(sql)