+++ init_ingester2.go
@@ -44,6 +44,8 @@ func initIngester2(app *App) {
 		storage.NewAssertPair(hRepo),
 		storage.NewPoll(hRepo),
 		storage.NewVote(hRepo),
+		storage.NewAccountSpecificRules(hRepo),
+		storage.NewSaleParticipation(hRepo),
 	)
 
 	idProvider := struct {